% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eraeroCan.r
\name{eraeroCan}
\alias{eraeroCan}
\title{Aerodynamic resistance terms}
\usage{
eraeroCan(u2, lai, fvc, h, lw)
}
\arguments{
\item{u2}{Wind speed at 2m height [m/s]}

\item{lai}{Leaf Area Index [m2/m2]}

\item{fvc}{Fractional vegetation (canopy) cover}

\item{h}{Canopy height [m]}

\item{lw}{Typical leaf width [m]}
}
\value{
Returns a vector of 3 aerodynamic resistance terms: above-canopy (raa), within-canopy (rca) and sub-canopy (rsa)
}
\description{
Aerodynamic resistance terms
}
