% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padstr (Automatisch gesichert).R, R/padstr.r
\name{pad1}
\alias{pad1}
\alias{pad1}
\title{Auxiliary function for processing of date/time strings}
\usage{
pad1(mon)

pad1(mon)
}
\description{
Auxiliary function for processing of date/time strings

Auxiliary function for processing of date/time strings
}
