% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padstr (Automatisch gesichert).R, R/padstr.r
\name{padstr}
\alias{padstr}
\alias{padstr}
\title{Auxiliary function for processing of date/time strings}
\usage{
padstr(mon)

padstr(mon)
}
\arguments{
\item{mon}{One- or two-digit number indicating day or month number}

\item{mon}{One- or two-digit number indicating day or month number}
}
\value{
A two-digit character string with a leading zero if the input was one digit

A two-digit character string with a leading zero if the input was one digit
}
\description{
Auxiliary function for processing of date/time strings

Auxiliary function for processing of date/time strings
}
