% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcJarvis2.r
\name{rcJarvis2}
\alias{rcJarvis2}
\title{Calculation of canopy resistance}
\usage{
rcJarvis2(laic, rsmin, rsmax, rg, sm, sfc, vp, vpsat, tAir, jR, jvpd, jsm, tl,
  th, topt, pa)
}
\arguments{
\item{laic}{Clumped LAI (LAI times FVC)}

\item{rsmin}{Minimum stomatal resistance [s/m]}

\item{rsmax}{Maximum stomatal resistance [s/m]}

\item{rg}{Global radiation [W/m2]}

\item{sm}{Current plant available soil moisture [mm]}

\item{sfc}{Size of plant available soil moisture reservoir [mm]}

\item{vp}{Atmospheric water vapor pressure [hPa]}

\item{vpsat}{Saturated vapor pressure [hPa]}

\item{tAir}{Air temperature [째C]}

\item{jR}{Parameter indicating stomatal sensitivity to radiation [W/m2]}

\item{jvpd}{Parameter indicating stomatal sensitivity to VPD [hPa-1]}

\item{jsm}{Parameter indicating stomatal sensitivity to relative extractable water [-]}

\item{tl}{Minimum temperature for photosynthesis [째C]}

\item{th}{Maximum temperature for photosynthesis [째C]}

\item{topt}{Optimal temperature for photosynthesis [째C]}

\item{pa}{Control parameter: Calculate potential transpiration if pa==1, and actual transpiration of pa==2}
}
\description{
Calculation of canopy resistance
}
