% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.forhytm.R
\name{run.forhytm}
\alias{run.forhytm}
\title{Model core of FORHYTM}
\usage{
run.forhytm(t, p, u, rh, rg, rssd, elv = 1190, sfc = 150, kappa = 15,
  laimax = 4, gsLen = 180, cBeta = 2, rsmin = 180, jvpd = 0.025,
  topt = 25, ff = 0.386, z = 25, ksimax = 2, jsm = 0.4, tl = 0,
  th = 40, extCoeff = 0.5, jR = 100, rsmax = 5000, alb = 0.1,
  lw = 0.01)
}
\arguments{
\item{t}{Vector containing hourly values of air temperature [째C]}

\item{p}{Vector containing hourly values of precipitation [mm]}

\item{u}{Vector containing hourly values of wind speed [m/s]}

\item{rh}{Vector containing hourly values of relative humidity (0...1)}

\item{rg}{Vector containing hourly values of global radiation [W/m2]}

\item{rssd}{Vector containing daily values of relative sunshine duration (0...1)}

\item{elv}{Elevation [m asl]}

\item{sfc}{Size of plant available soil moisture reservoir [mm]}

\item{kappa}{e-folding of soil evaporation reduction function [days]}

\item{laimax}{Leaf Area Index at full foliage cover [m2/m2]}

\item{gsLen}{Length of leafed season, from bud burst to leaf fall [days]}

\item{cBeta}{Shape parameter of soil moisture recharge [-]}

\item{rsmin}{Minimum stomatal resistance [s/m]}

\item{jvpd}{Parameter indicating stomatal sensitivity to VPD [hPa-1]}

\item{topt}{Optimal temperature for photosynthesis [째C]}

\item{ff}{Parameter linking fractional canopy cover to LAI}

\item{z}{Canopy height [m]}

\item{ksimax}{Parameter linking canopy interception capacity to LAI}

\item{jsm}{Parameter indicating stomatal sensitivity to relative extractable water [-]}

\item{tl}{Minimum temperature for photosynthesis [째C]}

\item{th}{Maximum temperature for photosynthesis [째C]}

\item{extCoeff}{Canopy light extinction coefficient}

\item{rsmax}{Maximum stomatal resistance [s/m]}

\item{alb}{albedo (0...1)}

\item{lw}{Typical leaf width [m]}

\item{Parameter}{indicating stomatal sensitivity to radiation [W/m2]}
}
\value{
Returns a data.frame containing daily values of water fluxes (potential transpiration, actual transpiration, soil evaporation, interception evaporation, water yield) and storage (canopy storage, soil moisture, snowpack).
}
\description{
Model core of FORHYTM
}
